[#829] Compose both composition roots in the unit suite
The host's registrations lived in top-level statements, which cannot be
called, so nothing built the real graph: a dependency nobody registered
surfaced as an exception out of a worker rather than as a suite that
failed. They move into HostComposition.Compose, which hands the request
pipeline the surfaces it decided; Program.cs keeps the bootstrap logger,
the provisioned configuration, the environment-only rejection, and the
pipeline.

HostCompositionTests builds a provider over that with ValidateOnBuild and
ValidateScopes on and then resolves every service the MailFathom
assemblies registered, because the container sees through a constructor
and not through a factory. It carries one shape per set of switches, so
the conditional half of the composition — the embedding chain, the chat
endpoint and its relevance filter, each scanner, the spam daemon, the
surfaces served — is composed somewhere other than an operator's server.

The command has no container, so its version of the same defect is a
command class nothing attached. CliRootCommandTests compares the tree
CliRootCommand.Create builds against the commands the assembly declares,
and refuses two siblings sharing a name.
