+++ tests/AGENTS.md
@@ -14,7 +14,8 @@ The **.NET and C# conventions** in `src/AGENTS.md` apply to test code as well. A
 - `Boundaries.UnitTests` is named after no production boundary because its subject is what holds between them, and it is the one project whose rules read compiled intermediate language rather than assembly metadata. Reach for it only where reflection structurally cannot answer the claim: a dependency that exists solely inside a method body, and a rule about a namespace within an assembly. Everything else stays where it is — `ApplicationDependencyBoundaryTests`, `McpDependencyBoundaryTests`, and `SecretBoundaryArchitectureTests` assert an *exact* reference set, which is the stronger assertion for that question because it puts a reference added for an ordinary reason in front of a reviewer instead of only refusing a forbidden one. Do not migrate them here and do not restate them here. A rule added here names the contract it holds in its `Because`, since that text is what a developer who trips it reads, and it names any type it has to admit rather than narrowing the rule until nothing is excluded. The suite is the slowest in the repository per assertion — loading eight assemblies' intermediate language costs about ten seconds once — so it holds a handful of rules about invariants worth that, not a rule per convention.
 - `SharedSources.UnitTests` is the exception, because it covers the two directories no project owns: `src/shared` and `tests/shared`. Both are compiled into their consumers through `Compile Include` items, so this project compiles them the same way a consumer does and asserts their contracts in one place. The name is `SharedSources`, not `Shared`, because `Shared` is a reserved word in another .NET language and CA1716 rejects it as a namespace.
 - Assert a shared marker attribute against a sample type declared beside its tests, not against whichever boundary applies it today. A marker compiled into several assemblies is a distinct type in each, so `IsDefined(typeof(...))` across an assembly boundary matches nothing and the assertion passes without inspecting anything. A local sample keeps the types identical; read the marker by name regardless, because that is how the coverage collector reads it.
-- `Host.UnitTests` covers what the composition root decides rather than how it wires: options validation rules, startup fail-fast behavior, hosted-service loops and their failure isolation, and the mapping from bound configuration onto infrastructure contracts. It does not build the real `WebApplication`, start a server, or assert dependency-injection registrations; the container is exercised only where a test needs a real scope. That belongs to the later integration phase.
+- `Host.UnitTests` covers what the composition root decides: options validation rules, startup fail-fast behavior, hosted-service loops and their failure isolation, and the mapping from bound configuration onto infrastructure contracts. It does not build the real `WebApplication` and does not start a server — the request pipeline is the composed host's to prove, and the integration suite is where that happens.
+- It does compose the real service graph, in `HostCompositionTests`, and that is the one place anything does. `HostComposition.Compose` is callable for exactly this reason, and the test builds a provider over it with `ValidateOnBuild` and `ValidateScopes` on and then resolves every service the MailFathom assemblies registered. The resolution is the point rather than the validation: the container sees through a constructor and not through a factory, most of what the composition root registers is a factory, and a service nobody registered behind one is otherwise first missed when a worker asks for it on somebody's server. Which registrations exist follows from configuration, so the test carries one shape per set of switches — an embedding chain, a chat endpoint and its relevance filter, each scanner, the spam daemon, the surfaces served — and a new switch that decides a registration takes a shape of its own in the same change. Nothing in it reaches a database, a socket, or the file system, and a registration that would is the finding rather than the test's problem to work around.
 - Put logic worth unit-testing in `Domain`, `Application`, or `Infrastructure` rather than in `Host`. `Host.UnitTests` exists because some host-owned decisions are real, not as a licence to grow the composition root.
 - Follow Arrange, Act, Assert. Add explicit `// Arrange`, `// Act`, and `// Assert` comments.
 - A test helper needed by more than one boundary lives in `tests/shared/` and is linked with a `Compile Include` item, the same arrangement `src/shared/` uses. Keep such a file free of production behavior and give it the assembly-neutral `MailFathom.TestSupport` namespace.
